TNT_BRAIN/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
tavily-python
pydantic
python-multipart
openai
tiktoken